requirements.txt
openai==1.63.2
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
itsdangerous==2.2.0